+++ djangojs/templates/djangojs/init.js
@@ -1,0 +1,4 @@
+(function(){
+window.DJANGO_JS_URLS = {{ urls|safe }};
+window.DJANGO_JS_CONTEXT = {{ context|safe }};
+}());