+++ djangojs/urls.py
@@ -6,7 +6,7 @@ from os.path import join, isdir
 from django.conf.urls import patterns, url
 
 from djangojs.conf import settings
-from djangojs.views import UrlsJsonView, ContextJsonView
+from djangojs.views import UrlsJsonView, ContextJsonView, JsInitView
 
 js_info_dict = {
     'packages': [],
@@ -19,6 +19,7 @@ for app in settings.INSTALLED_APPS:
             js_info_dict['packages'].append(app)
 
 urlpatterns = patterns('',
+    url(r'^init\.js$', JsInitView.as_view(), name='django_js_init'),
     url(r'^urls$', UrlsJsonView.as_view(), name='django_js_urls'),
     url(r'^context$', ContextJsonView.as_view(), name='django_js_context'),
     url(r'^translation$', 'django.views.i18n.javascript_catalog', js_info_dict, name='js_catalog'),